Refactor logic

+++ PharmaShop-server/pharmacy.Api/Controllers/DiscountController.cs
@@ -1,6 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Mvc;
 using pharmacy.Api.Responses;
 using pharmacy.Core;
+using pharmacy.Core.DTOs.Discount;
 using pharmacy.Core.Entities;
 namespace pharmacy.Api.Controllers;
 
@@ -10,26 +12,31 @@ public class DiscountController : ControllerBase
 {
     private readonly IUnitOfWork _unitOfWork;
     private readonly IResponseHandler _responseHandler;
+    private readonly IMapper _mapper; 
 
-    public DiscountController(IUnitOfWork unitOfWork, IResponseHandler responseHandler)
+    public DiscountController(IUnitOfWork unitOfWork, IResponseHandler responseHandler, IMapper mapper)
     {
         _unitOfWork = unitOfWork;
         _responseHandler = responseHandler;
+        _mapper = mapper; 
     }
 
-    [HttpGet("top/{topN}")]
-    public async Task<IActionResult> GetTopDiscountsAsync(int topN)
+    [HttpPost]
+    public async Task<IActionResult> CreateDiscountAsync([FromBody] DiscountRequestDto discountRequestDto)
     {
         try
         {
-            var discounts = await _unitOfWork.discountRepository.GetTopDiscountsAsync(topN, DateTime.UtcNow);
-
-            if (discounts == null || discounts.Count() == 0)
+            if (discountRequestDto == null)
             {
-                return _responseHandler.NotFound("No active discounts found.");
+                return _responseHandler.BadRequest("Invalid discount data.");
             }
 
-            return _responseHandler.Success(discounts, "Top discounts fetched successfully.");
+            var discount = _mapper.Map<Discount>(discountRequestDto);
+
+            var createdDiscount = await _unitOfWork.discountRepository.CreateAsync(discount);
+            _unitOfWork.Complete();
+
+            return _responseHandler.Created(createdDiscount, "Discount created successfully.");
         }
         catch (Exception ex)
         {
@@ -46,28 +53,30 @@ public class DiscountController : ControllerBase
 
             if (discount == null)
                 return _responseHandler.NotFound($"Discount with ID {id} not found.");
+            var discountResponseDto = _mapper.Map<DiscountResponseDto>(discount);
 
-            return _responseHandler.Success(discount, "Discount details fetched successfully.");
+            return _responseHandler.Success(discountResponseDto, "Discount details fetched successfully.");
         }
         catch (Exception ex)
         {
             return _responseHandler.BadRequest($"An error occurred: {ex.Message}");
         }
     }
-
-    [HttpPost]
-    public async Task<IActionResult> CreateDiscountAsync([FromBody] Discount discount)
+    [HttpGet]
+    public async Task<IActionResult> GetAllDiscountsAsync()
     {
         try
         {
-            if (discount is null)
+            var discounts = await _unitOfWork.discountRepository.GetAllAsync();
+
+            if (discounts == null || !discounts.Any())
             {
-                return _responseHandler.BadRequest("Invalid discount data.");
+                return _responseHandler.NotFound("No discounts found.");
             }
 
-            var createdDiscount = await _unitOfWork.discountRepository.CreateAsync(discount);
+            var discountResponseDtos = _mapper.Map<IEnumerable<DiscountResponseDto>>(discounts);
 
-            return _responseHandler.Created(createdDiscount, "Discount created successfully.");
+            return _responseHandler.Success(discountResponseDtos, "Discounts fetched successfully.");
         }
         catch (Exception ex)
         {
@@ -76,23 +85,18 @@ public class DiscountController : ControllerBase
     }
 
     [HttpPut("{id}")]
-    public async Task<IActionResult> UpdateDiscountAsync(int id, [FromBody] Discount discount)
+    public async Task<IActionResult> UpdateDiscountAsync(int id, [FromBody] DiscountRequestDto discountRequestDto)
     {
         try
         {
-            if (discount == null || discount.DiscountId != id)
-            {
-                return _responseHandler.BadRequest("Invalid discount data or ID mismatch.");
-            }
+            var existingDiscount = await _unitOfWork.discountRepository.GetByID(id);
 
-            var updatedDiscount = await _unitOfWork.discountRepository.UpdateAsync(id, discount);
+            var updatedDiscount = _mapper.Map(discountRequestDto, existingDiscount);
 
-            if (updatedDiscount == null)
-            {
-                return _responseHandler.NotFound($"Discount with ID {id} not found.");
-            }
+            var result = await _unitOfWork.discountRepository.UpdateAsync(id, updatedDiscount);
+            _unitOfWork.Complete();
 
-            return _responseHandler.Success(updatedDiscount, "Discount updated successfully.");
+            return _responseHandler.Success(result, "Discount updated successfully.");
         }
         catch (Exception ex)
         {
@@ -105,7 +109,15 @@ public class DiscountController : ControllerBase
     {
         try
         {
+            var existingDiscount = await _unitOfWork.discountRepository.GetByID(id);
+
+            if (existingDiscount == null)
+            {
+                return _responseHandler.NotFound($"Discount with ID {id} not found.");
+            }
+
             var deleted = await _unitOfWork.discountRepository.DeleteAsync(id);
+            _unitOfWork.Complete();
 
             return _responseHandler.NoContent("Discount deleted successfully.");
         }
@@ -114,4 +126,7 @@ public class DiscountController : ControllerBase
             return _responseHandler.BadRequest($"An error occurred: {ex.Message}");
         }
     }
+
+
+
 }
